Examen Casi Listo

+++ Segundo Parcial/ENTIDADES/RegistrodeArticulos.cs
@@ -13,7 +13,7 @@ namespace Segundo_Parcial.ENTIDADES
         public int ArticulosId { get; set; }
         public string Descripcion { get; set; }
         public decimal costo { get; set; }
-        public int Ganancia { get; set; }
+        public decimal Ganancia { get; set; }
         public decimal precio { get; set; }
         public int Inventario { get; set; }
